refactor: move tests to per-command test files
